requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
discord.py==1.4.1
distlib==0.3.1
filelock==3.0.12
idna==2.10
iniconfig==1.0.1
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
multidict==4.7.6
packaging==20.4
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pypresence==4.1.2
pytest==6.0.2
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.34
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
